Feature/product crud api

+++ src/main/java/com/goenaga/shop/security/service/JWTService.java
@@ -15,6 +15,7 @@ import javax.crypto.SecretKey;
 import java.util.Base64;
 import java.util.Date;
 import java.util.Objects;
+import java.util.Optional;
 
 
 @Service
@@ -28,20 +29,26 @@ public class JWTService {
     private final TokenRepository tokenRepository;
 
     public String createToken(User user) {
-//        Delete previous issued tokens to the user
-        tokenRepository.deleteByEmail(user.getEmail());
+//        Delete previous issued token to the user if it has one assigned
+        if (tokenRepository.findTokenByEmail(user.getEmail()).isPresent()) {
+            tokenRepository.deleteByEmail(user.getEmail());
+        }
 
 //        Issue new token
         String jwtToken = issueToken(user.getEmail());
-        tokenRepository.save(TokenEntity.builder().token(jwtToken).email(user.getEmail()).build());
+        tokenRepository.save(TokenEntity.builder().email(user.getEmail()).token(jwtToken).build());
         return TOKEN_PREFIX + jwtToken;
     }
 
     public String refreshToken(String token) throws ServletException {
 //        Authenticate token and retrieve email
         String email = parseClaims(token).getSubject();
-//        Check if expired token is in DB.
-        String lastIssuedToken = tokenRepository.findTokenByEmail(email).getToken();
+
+//        Check if an expired token exists in the DB for the user.
+        String lastIssuedToken = null;
+        Optional<TokenEntity> tokenInDB = tokenRepository.findTokenByEmail(email);
+        if (tokenInDB.isPresent()) { lastIssuedToken = tokenInDB.get().getToken(); }
+
 
 //        If provided expired token equals last issued token, issue a new token
         if (Objects.equals(token, lastIssuedToken)) {
